refactor(web): convert member-detail drawer to a centered modal
Replace the hand-rolled right-side drawer on the Members page with a centered
native <dialog> modal (matching ConfirmModal / BulkActionsBar), gaining
focus-trap, Escape, and an inert backdrop for free, and unifying every close
path through one handler.

Extract the presentation helpers shared by the list rows and the detail view
(initialsFor, GitHubIdentity, ClassificationBadge, runInviteMember) into
memberPresentation.tsx, and move the detail view into MemberDetailModal.tsx.
OrgMembersPage drops ~400 lines and its now-unused imports; behavior is
unchanged.
